refactor and included new routes for reconciler and query
