+++ .idea/libraries/lib.xml
@@ -1,0 +1,25 @@
+<component name="libraryTable">
+  <library name="lib">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/hk2-locator-2.4.0-b09.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/hk2-utils-2.4.0-b09.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/javax.ws.rs-api-2.0.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/json-20170516.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/javax.inject-2.4.0-b09.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/hk2-api-2.4.0-b09.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/osgi-resource-locator-1.0.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/jersey-client-2.16.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/jersey-container-servlet-core-2.16.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/javassist-3.18.1-GA.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/jersey-guava-2.16.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/jersey-server-2.16.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/javax.annotation-api-1.2.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/aopalliance-repackaged-2.4.0-b09.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/jersey-media-jaxb-2.16.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/jersey-common-2.16.jar!/" />
+      <root url="jar://$PROJECT_DIR$/TripleDataProcessor/target/TripleDataProcessor/WEB-INF/lib/validation-api-1.1.0.Final.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
