+++ TripleDataProcessor/src/main/java/org/ECS193/TripleDataProcessor/data/ObjectElement.java
@@ -1,4 +1,4 @@
-package org.ECS193.TripleDataProcessor;
+package org.ECS193.TripleDataProcessor.data;
 
 public class ObjectElement extends TripleElement{
