+++ TripleDataProcessor/src/main/java/org/ECS193/TripleDataProcessor/data/PredicateElement.java
@@ -1,4 +1,4 @@
-package org.ECS193.TripleDataProcessor;
+package org.ECS193.TripleDataProcessor.data;
 
 public class PredicateElement extends TripleElement{
 	public PredicateElement(String name, TYPE type) {
