+++ TripleDataProcessor/src/main/java/org/ECS193/TripleDataProcessor/data/TripleElement.java
@@ -1,4 +1,4 @@
-package org.ECS193.TripleDataProcessor;
+package org.ECS193.TripleDataProcessor.data;
 
 public abstract class TripleElement {
 	private String name;
