+++ TripleDataProcessor/src/main/java/org/ECS193/TripleDataProcessor/routes/Reconcile.java
@@ -1,0 +1,19 @@
+package org.ECS193.TripleDataProcessor.routes;
+
+import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+
+@Path("reconcile")
+public class Reconcile {
+
+	@POST
+	public Response post() {
+		return null;
+	}
+
+}
